+++ Hypixel.NET/PlayerApi/Player/Stats/SuperSmash.cs
@@ -65,7 +65,7 @@ namespace Hypixel.NET.PlayerApi.Player.Stats
         public int DamageDealtNormal { get; set; }
 
         [JsonProperty("xp_THE_BULK")]
-        public double XpTheBulk { get; set; }
+        public int XpTheBulk { get; set; }
 
         [JsonProperty("smash_level_total")]
         public int SmashLevelTotal { get; set; }
@@ -104,7 +104,7 @@ namespace Hypixel.NET.PlayerApi.Player.Stats
         public int FriendLossesNormal { get; set; }
 
         [JsonProperty("xp_CAKE_MONSTER")]
-        public double XpCakeMonster { get; set; }
+        public int XpCakeMonster { get; set; }
 
         [JsonProperty("lastLevel_CAKE_MONSTER")]
         public int LastLevelCakeMonster { get; set; }
@@ -125,7 +125,7 @@ namespace Hypixel.NET.PlayerApi.Player.Stats
         public int DamageDealt2V2 { get; set; }
 
         [JsonProperty("xp_GENERAL_CLUCK")]
-        public double XpGeneralCluck { get; set; }
+        public int XpGeneralCluck { get; set; }
 
         [JsonProperty("lastLevel_GENERAL_CLUCK")]
         public int LastLevelGeneralCluck { get; set; }
@@ -140,7 +140,7 @@ namespace Hypixel.NET.PlayerApi.Player.Stats
         public int Losses2V2 { get; set; }
 
         [JsonProperty("xp_BOTMUN")]
-        public double XpBotmun { get; set; }
+        public int XpBotmun { get; set; }
 
         [JsonProperty("lastLevel_BOTMUN")]
         public int LastLevelBotmun { get; set; }
@@ -170,13 +170,13 @@ namespace Hypixel.NET.PlayerApi.Player.Stats
         public int LossesTeams { get; set; }
 
         [JsonProperty("xp_FROSTY")]
-        public double XpFrosty { get; set; }
+        public int XpFrosty { get; set; }
 
         [JsonProperty("lastLevel_FROSTY")]
         public int LastLevelFrosty { get; set; }
 
         [JsonProperty("xp_TINMAN")]
-        public double XpTinman { get; set; }
+        public int XpTinman { get; set; }
 
         [JsonProperty("lastLevel_TINMAN")]
         public int LastLevelTinman { get; set; }
@@ -185,22 +185,22 @@ namespace Hypixel.NET.PlayerApi.Player.Stats
         public int LastLevelMarauder { get; set; }
 
         [JsonProperty("xp_MARAUDER")]
-        public double XpMarauder { get; set; }
+        public int XpMarauder { get; set; }
 
         [JsonProperty("xp_GOKU")]
-        public double XpGoku { get; set; }
+        public int XpGoku { get; set; }
 
         [JsonProperty("lastLevel_GOKU")]
         public int LastLevelGoku { get; set; }
 
         [JsonProperty("xp_SPODERMAN")]
-        public double XpSpoderman { get; set; }
+        public int XpSpoderman { get; set; }
 
         [JsonProperty("lastLevel_SPODERMAN")]
         public int LastLevelSpoderman { get; set; }
 
         [JsonProperty("xp_PUG")]
-        public double XpPug { get; set; }
+        public int XpPug { get; set; }
 
         [JsonProperty("lastLevel_PUG")]
         public int LastLevelPug { get; set; }
@@ -212,7 +212,7 @@ namespace Hypixel.NET.PlayerApi.Player.Stats
         public int VotesGunmetal { get; set; }
 
         [JsonProperty("xp_SKULLFIRE")]
-        public double XpSkullfire { get; set; }
+        public int XpSkullfire { get; set; }
 
         [JsonProperty("lastLevel_SKULLFIRE")]
         public int LastLevelSkullfire { get; set; }
@@ -227,7 +227,7 @@ namespace Hypixel.NET.PlayerApi.Player.Stats
         public bool MasterArmorFrosty { get; set; }
 
         [JsonProperty("xp_SHOOP_DA_WHOOP")]
-        public double XpShoopDaWhoop { get; set; }
+        public int XpShoopDaWhoop { get; set; }
 
         [JsonProperty("lastLevel_SHOOP_DA_WHOOP")]
         public int LastLevelShoopDaWhoop { get; set; }
@@ -245,7 +245,7 @@ namespace Hypixel.NET.PlayerApi.Player.Stats
         public bool MasterArmorTheBulk { get; set; }
 
         [JsonProperty("xp_DUSK_CRAWLER")]
-        public double XpDuskCrawler { get; set; }
+        public int XpDuskCrawler { get; set; }
 
         [JsonProperty("lastLevel_DUSK_CRAWLER")]
         public int LastLevelDuskCrawler { get; set; }
@@ -281,16 +281,16 @@ namespace Hypixel.NET.PlayerApi.Player.Stats
         public int LastLevelSergeantShield { get; set; }
 
         [JsonProperty("xp_SERGEANT_SHIELD")]
-        public double XpSergeantShield { get; set; }
+        public int XpSergeantShield { get; set; }
 
         [JsonProperty("xp_SANIC")]
-        public double XpSanic { get; set; }
+        public int XpSanic { get; set; }
 
         [JsonProperty("lastLevel_SANIC")]
         public int LastLevelSanic { get; set; }
 
         [JsonProperty("xp_GREEN_HOOD")]
-        public double XpGreenHood { get; set; }
+        public int XpGreenHood { get; set; }
 
         [JsonProperty("lastLevel_GREEN_HOOD")]
         public int LastLevelGreenHood { get; set; }
